+++ InquirerPy/prompts/__init__.py
@@ -6,5 +6,6 @@ from InquirerPy.prompts.filepath import FilePathPrompt
 from InquirerPy.prompts.fuzzy import FuzzyPrompt
 from InquirerPy.prompts.input import InputPrompt
 from InquirerPy.prompts.list import ListPrompt
+from InquirerPy.prompts.number import NumberPrompt
 from InquirerPy.prompts.rawlist import RawlistPrompt
 from InquirerPy.prompts.secret import SecretPrompt