+++ examples/alternate/number.py
@@ -1,0 +1,3 @@
+from InquirerPy import inquirer
+
+# result = inquirer.